Rename module emojies to emojis.
