+++ src/pipx/emojis.py
